backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
Django==3.2.9
django-cors-headers==3.10.0
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
drf-nested-routers==0.93.4
ecdsa==0.17.0
idna==3.3
josepy==1.10.0
mozilla-django-oidc==2.0.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
pyasn1==0.4.8
pycparser==2.20
pyOpenSSL==21.0.0
python-jose==3.3.0
python-keycloak==0.25.0
pytz==2021.3
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
treelib==1.6.1
datamodel-code-generator==0.11.15
